Have libexec scripts be able to communicate with a docker container
start-target creates and starts a docker container from the image created
by make-base-vm

stop-target stops and removes the docker container from start-target

copy-from-target, copy-to-target, and on-target can execute commands
and copy files to and from the docker container started by start-target.

make-clean-vm does nothing because nothing needs to be done to make a
clean docker container.